work in progress, but still fonctional, delete of independant bone half done
git-svn-id: http://www.genecraft-project.com/svn/genecraft/trunk@137 2ce2bb41-2ff2-4fd5-96a4-e3c75a7bd90e

+++ src/genelab/core/entities/families/antfamily.cpp
@@ -19,7 +19,7 @@ AntFamily::AntFamily() : EntityFamily()
     // head / trunk / metasoma
     headRadius      = Tools::random(0.5,1.2);
     trunkRadius     = Tools::random(0.8,1.2);
-    metasomaRadius  = Tools::random(headRadius,headRadius+1.0);
+    metasomaRadius  = Tools::random((double)headRadius, (double)headRadius+1.0);
 
     // legs
     nbBonesInLeg    = Tools::random(2,3);
